Small modifications of affectation.h due to test under Windows plateform
git-svn-id: svn://scm.gforge.inria.fr/svnroot/paradiseo@1068 331e1502-861f-0410-8da2-ba01fb791d7f

+++ trunk/paradiseo-mo/tutorial/examples/function/affectation.h
@@ -40,8 +40,8 @@
 #include <eoScalarFitness.h>
 #include <EO.h>
 
-// A float that has to be minimized.
-typedef eoScalarFitness< float, std::greater<double> > functionFitness ;
+// A double that has to be minimized.
+typedef eoScalarFitness< double, std::greater<double> > functionFitness ;
 
 class Affectation : public EO<functionFitness>, public std::pair<double, double>
 {};
